+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [module:quote] improve new user quote interface
 * [module:quote] expose client infos to invoice templates
+* [Services] Remove deprecated functions from servicesController
 * [Exceptions] Add PfmUserException class
 * [Users] Remove login from edit function
 * [self_registration] fix selfregistration email sending order